+++ commands/clear.js
@@ -1,3 +1,4 @@
+const { Permissions } = require("discord.js");
 const utils = require('../utils.js');
 
 let settings = {
@@ -18,8 +19,17 @@ module.exports = {
     args: settings.args,
     usage: settings.usage,
     execute: async(message, args) => {
+        if (!message.member.permissionsIn(message.channel).has(Permissions.FLAGS.MANAGE_MESSAGES)) {
+            console.log(`User ${message.author} tried clearing messages in channel ${message.channel}`);
+            return;
+        }
+
         if (isNaN(args[0]) || parseInt(args[0]) <= 0) {
             return message.reply(utils.getHelpMessage(message.client, message.client.commands.get(settings.name)));
+        }
+
+        if (!message.guild.me.permissionsIn(message.channel).has(Permissions.FLAGS.MANAGE_MESSAGES)) {
+            return message.reply(`Je n'ai pas la permission de supprimmer des messages ici :pensive:`);
         }
 
         let toDelete = parseInt(args[0]) > 99 ? 100 : parseInt(args[0]) + 1;